+++ src/gtk4_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: 1dbb8dd6ff9d9201c8b0d8ea9b9ef85834c45644 */
+ * Stub hash: 51bb6da828e9e01c7327be4fc3ed530f7d8308ef */
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_class_Gtk4_GObject___construct, 0, 0, 0)
 ZEND_END_ARG_INFO()
@@ -119,6 +119,12 @@ ZEND_END_ARG_INFO()
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Gtk4_GLib_main_context_iteration, 0, 0, _IS_BOOL, 0)
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, may_block, _IS_BOOL, 0, "false")
 ZEND_END_ARG_INFO()
+
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Gtk4_GLib_set_prgname, 0, 1, IS_VOID, 0)
+	ZEND_ARG_TYPE_INFO(0, prgname, IS_STRING, 0)
+ZEND_END_ARG_INFO()
+
+#define arginfo_class_Gtk4_GLib_get_prgname arginfo_class_Gtk4_GParamSpec_get_nick
 
 #define arginfo_class_Gtk4_GMainLoop___construct arginfo_class_Gtk4_GObject___construct
 
@@ -449,6 +455,8 @@ ZEND_METHOD(Gtk4_GLib, timeout_add);
 ZEND_METHOD(Gtk4_GLib, io_add_watch);
 ZEND_METHOD(Gtk4_GLib, source_remove);
 ZEND_METHOD(Gtk4_GLib, main_context_iteration);
+ZEND_METHOD(Gtk4_GLib, set_prgname);
+ZEND_METHOD(Gtk4_GLib, get_prgname);
 ZEND_METHOD(Gtk4_GMainLoop, __construct);
 ZEND_METHOD(Gtk4_GMainLoop, run);
 ZEND_METHOD(Gtk4_GMainLoop, quit);
@@ -590,6 +598,8 @@ static const zend_function_entry class_Gtk4_GLib_methods[] = {
 	ZEND_ME(Gtk4_GLib, io_add_watch, arginfo_class_Gtk4_GLib_io_add_watch, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
 	ZEND_ME(Gtk4_GLib, source_remove, arginfo_class_Gtk4_GLib_source_remove, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
 	ZEND_ME(Gtk4_GLib, main_context_iteration, arginfo_class_Gtk4_GLib_main_context_iteration, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
+	ZEND_ME(Gtk4_GLib, set_prgname, arginfo_class_Gtk4_GLib_set_prgname, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
+	ZEND_ME(Gtk4_GLib, get_prgname, arginfo_class_Gtk4_GLib_get_prgname, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
 	ZEND_FE_END
 };